ai-career-mentor-final/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx
pydantic
pymongo
python-jose
passlib[bcrypt]
openai
spacy
sentence-transformers
